requirements.txt
asgiref==3.8.1
better-profanity==0.7.0
black==24.10.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.8
distlib==0.3.9
Django==5.1.4
django-environ==0.11.2
django-parler==2.3
django-rosetta==0.10.1
django-stubs==5.1.1
django-stubs-ext==5.1.1
filelock==3.16.1
flake8==7.1.1
identify==2.6.3
idna==3.10
isort==5.13.2
mccabe==0.7.0
mypy==1.14.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
polib==1.2.0
pre_commit==4.0.1
pycodestyle==2.12.1
pyflakes==3.2.0
pytils==0.4.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
types-PyYAML==6.0.12.20241221
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.28.0
